bitti

+++ ft_printf.h
@@ -15,17 +15,16 @@
 
 # include <stdarg.h>
 # include <unistd.h>
-# include <stdlib.h>
 
 int	ft_printf(const char *format, ...);
 
 int	ft_putchar(char c);
 int	ft_putstr(char *str);
 
-int	ft_putnbr(int n);
-int	ft_putnbr_unsigned(unsigned int n);
+int	ft_putnbr(long n);
+int	ft_putnbr_unsigned(unsigned long n);
 
-int	ft_puthex(unsigned int n, char format);
+int	ft_puthex(unsigned long n, char format);
 int	ft_putpointer(unsigned long ptr);
 
 #endif
